applications/edge-llm-service/requirements.txt
flask==2.2.3
gunicorn==20.1.0
transformers==4.28.1
numpy==1.24.2
requests==2.28.2
python-dotenv==1.0.0
tensorflow-lite==2.12.0
onnxruntime==1.14.1
prometheus-flask-exporter==0.22.3 